move ShapeError to exceptions.py
