+++ pymc3/exceptions.py
@@ -3,6 +3,7 @@ __all__ = [
     "IncorrectArgumentsError",
     "TraceDirectoryError",
     "ImputationWarning",
+    "ShapeError"
 ]
 
 
@@ -24,3 +25,12 @@ class ImputationWarning(UserWarning):
     """Warning that there are missing values that will be imputed."""
 
     pass
+
+
+class ShapeError(Exception):
+    """Error that the shape of a variable is incorrect."""
+    def __init__(self, message, actual=None, expected=None):
+        if expected and actual:
+            super().__init__('{} (actual {} != expected {})'.format(message, actual, expected))
+        else:
+            super().__init__(message)
