+++ pymc3/ode/__init__.py
@@ -1,2 +1,2 @@
 from . import utils
-from .ode import DifferentialEquation, ShapeError
+from .ode import DifferentialEquation
